requirements.txt
openai
elevenlabs
networkx
matplotlib
moviepy
Flask==3.0.2
Flask-Cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
dnspython==2.6.1
sentence-transformers

